# Reorder lexer switch statement for perf boost.